+++ src/main/java/com/joe/coffee/api/Exception/ArtisanTorrefacteurExceptions/ArtisanTorrefacteurNotFoundException.java
@@ -2,7 +2,7 @@ package com.joe.coffee.api.Exception.ArtisanTorrefacteurExceptions;
 
 public class ArtisanTorrefacteurNotFoundException extends RuntimeException {
 
-    public ArtisanTorrefacteurNotFoundException(Long id) {
+    public ArtisanTorrefacteurNotFoundException(Integer id) {
 
         super("Artisan Torréfacteur avec l'id " + id + " introuvable.");
