+++ src/main/java/com/joe/coffee/api/Exception/DistributeurExceptions/DistributeurNotFoundException.java
@@ -2,7 +2,7 @@ package com.joe.coffee.api.Exception.DistributeurExceptions;
 
 public class DistributeurNotFoundException extends RuntimeException {
 
-    public DistributeurNotFoundException(Long id) {
+    public DistributeurNotFoundException(Integer id) {
 
         super("Distributeur avec l'id " + id + " introuvable.");
